requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
peewee==3.17.0
peewee-migrate==1.12.2
chromadb==0.4.22
numpy<2.0.0
huggingface_hub>=0.20.0,<0.25.0
pytesseract==0.3.10
PyMuPDF==1.23.8
sentence-transformers>=2.7.0
jinja2==3.1.3
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.4
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
itsdangerous==2.1.2
matplotlib==3.7.2